Implemented a Minimum Dot
Allows control of how strait forward the check for the player is. When 1 the dot will basically wait for the player to be in a direct line in front of it, when 0 will check ahead of the sides. Hope this explains it a bit, maybe you would understand better looking how I did it.